#tf-data-service Fix a type mismatch.
Diagnostics generated by clang-tidy: deprecated declarations

'int64' is deprecated: Use int64_t instead.

PiperOrigin-RevId: 448622987

+++ tensorflow/core/data/service/dispatcher_impl.cc
@@ -234,7 +234,7 @@ Status DataServiceDispatcherImpl::Start() {
 
 size_t DataServiceDispatcherImpl::NumActiveJobs() TF_LOCKS_EXCLUDED(mu_) {
   mutex_lock l(mu_);
-  int64 count = 0;
+  size_t count = 0;
   for (const auto& job : state_.ListJobs()) {
     if (!job->finished) {
       count++;
